Fixed event routing and moved sampleData into testing directory (and linked them)
