+++ app.js
@@ -5,6 +5,7 @@ const eventRoutes = require('./routes/eventRoutes');
 const userRoutes = require('./routes/userRoutes');
 const MongoStore = require('connect-mongo');
 const bodyParser = require('body-parser'); // For form data parsing
+const events = require('./testing/sampleData'); //for linking to sample data
 const app = express();
 require('dotenv').config(); //links .env file
 
@@ -21,26 +22,11 @@ mongoose.connect(url, { useNewUrlParser: true, useUnifiedTopology: true})
     });
 
 
-
-
 //Will be removed in future iteration
   // In-memory user for authentication demonstration
 const user = { email: 'user@example.com', password: 'password123' };
 const users = []; // Simple in-memory array to store new users
 
-// Event Data (for demonstration)
-const events = [
-  { id: 1, name: 'UNCC Campus Concert', date: '2024-09-25', time: '6:00 PM', location: 'UNC Charlotte Campus', description: 'Join us for an unforgettable concert at the Student Union!', admission: '20', imagePath: '' },
-  { id: 2, name: 'PNC Music Festival', date: '2024-09-30', time: '5:00 PM', location: 'PNC Music Pavilion', description: 'A weekend-long festival featuring popular artists from around the world!', admission: '50', imagePath: '/images/pnc_pavillion.jpg' },
-  { id: 3, name: 'Charlotte Car Show', date: '2024-10-05', time: '10:00 AM', location: 'Charlotte Convention Center', description: 'Check out the latest car models and meet fellow car enthusiasts!', admission: '15', imagePath: '' },
-  { id: 4, name: 'Food Truck Friday', date: '2024-10-10', time: '12:00 PM', location: 'Uptown Charlotte', description: 'Enjoy the best food trucks in Charlotte, featuring local delicacies and international flavors!', admission: '10', imagePath: '' },
-  { id: 5, name: 'UNCC Career Fair', date: '2024-10-15', time: '11:00 AM', location: 'UNC Charlotte Student Union', description: 'Meet recruiters from top companies and find your dream job!', admission: 'Free', imagePath: '' },
-  { id: 6, name: 'Carolina Panthers vs Falcons', date: '2024-10-20', time: '1:00 PM', location: 'Bank of America Stadium', description: 'Watch the Panthers take on the Falcons in an exciting NFL matchup!', admission: '100', imagePath: '/images/nfl_stadium.jpg' },
-  { id: 7, name: 'Latin Dance Night', date: '2024-10-22', time: '8:00 PM', location: 'The Fillmore Charlotte', description: 'An evening of salsa and bachata dancing with live music!', admission: '25', imagePath: '/images/uncc_latinX.jpg' },
-  { id: 8, name: 'Craft Beer Festival', date: '2024-10-28', time: '3:00 PM', location: 'BB&T Ballpark', description: 'Sample local craft beers from Charlotte\'s best breweries!', admission: '40', imagePath: '' },
-  { id: 9, name: 'UNCC Homecoming Parade', date: '2024-11-05', time: '12:00 PM', location: 'UNC Charlotte Campus', description: 'Celebrate UNCC Homecoming with a parade through campus!', admission: 'Free', imagePath: '' },
-  { id: 10, name: 'Winter Wonderland at Freedom Park', date: '2024-12-01', time: '5:00 PM', location: 'Freedom Park', description: 'A magical winter festival with ice skating, lights, and holiday treats!', admission: '25', imagePath: '' }
-];
 
 // Middleware to parse form data
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -64,10 +50,3 @@ app.use('/users', userRoutes);
 
 // Events Page route
 app.use('/events', eventRoutes);
-
-
-// Start the server
-/*const PORT = process.env.PORT || 3000;
-app.listen(PORT, () => {
-  console.log(`Server is running on http://localhost:${PORT}`);
-});*/
